Refactor public order endpoint into dedicated router
